+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [UNRELEASED]
 (nothing yet)
 
+## [0.4.2] - 2026-01-27
+### Changed
+- types now allows to define incoming messages for unknown events (it's breaking if you ignored exported consts and
+hardcoded names instead)
+
 ## [0.4.1] - 2026-01-25
 ### Changed
 - added OnCallback utility type